Get accepted variables

Get a data.frame with accepted variable names

accepted_variables(include=NULL, exclude=NULL)

Arguments

include

character or NULL. Can be used to use a subset of variable groups

exclude

character or NULL. Can be used to exclude variable groups

Value

data.frame

Examples

set_vocabulary("github:controvoc/terminag")
v <- accepted_variables("crop")
head(v)
#>    group         name      type unit required vocabulary multiple_allowed
#> 8   crop     trial_id character           yes                          no
#> 9   crop      plot_id character            no                          no
#> 10  crop     block_id character            no                          no
#> 11  crop growth_stage character            no                          no
#> 12  crop     LTE_name character            no                          no
#> 13  crop       season character            no     season               no
#>    valid_min valid_max NAok
#> 8         NA        NA  yes
#> 9         NA        NA  yes
#> 10        NA        NA  yes
#> 11        NA        NA  yes
#> 12        NA        NA  yes
#> 13        NA        NA  yes
#>                                                                                               description
#> 8  Identifier of a single trial in a single location and season (that may have treatments and replicates)
#> 9                           Identifier of a single plot in a single trial in a single location and season
#> 10                         Identifier of a single block in a single trial in a single location and season
#> 11                                                                        Growth stage description (code)
#> 12                                                                           Name of long-term experiment
#> 13                                                                                            Season name
#>                                  notes obo
#> 8   should be unique within dataset_id    
#> 9                                         
#> 10                                        
#> 11                                        
#> 12                                        
#> 13