Functions for compliance checking

Functions for compliance checking. You can check whether variable names are accepted or not, and if required variables are avaialble. You can also check whether character values are accepted and weather other variables are within specficied ranges. Furthermore you can check whether coordinates are in the country specified; whether dates are valid or whether a (date) span is within an expected limit.

check_variables(x, trms, required=TRUE, suggest=TRUE)
check_values(x, trms)
check_lonlat(x)
check_caps(x, vars, minchar=5, frac=0)
check_span(x, start, end, smin=0, smax=366)
check_date(x, name, trms=NULL)
check_datespan(x, startdate, enddate, smin=0, smax=366)

Arguments

x

data.frame to check

trms

data.frame with the vocabulary. E.g. accepted_values("github:controvoc/terminag")

required

logical. If TRUE the presence of required variables is checked

suggest

logical. If TRUE, unknown variable names are reported with a close match from the vocabulary (Levenshtein distance, as in carobiner::draft)

name

character. The variable name. For example, "planting_date"

vars

character. The variables to check

minchar

positive integer

frac

numeric

start

numeric

end

numeric

smin

numeric

smax

numeric

startdate

Date

enddate

Date

Value

data.frame

Examples

set_vocabulary("github:controvoc/terminag")
trms <- vocal::accepted_variables(exclude=c("metadata", "carob-metadata"))

x <- data.frame(
  dataset_id = 1,
  country="Côte d'Ivoire",
  location = "Taï",
  admin1 = "Montagnes",
  longitude=10,
  lat=12,
  longitude = -7.6,
  latitude = 5.875,
  yield = 3969,
  yield_part = "roots",
  crop = "manioc", # should be "cassava"
  planting_date = "2030-01-01",
  harvest_date = "2035-01-01",
  N_fertilizer = 100,
  P_fertilizer = 25,
  K_fertilize = -10,
  irrigated = FALSE,
  test = 0
)

check_variables(x, trms, required=FALSE, suggest=FALSE)
#>               check                                         msg
#> 1 unknown variables admin1, lat, longitude.1, K_fertilize, test

check_variables(x, trms, required=FALSE, suggest=TRUE)
#>               check                         msg
#> 1  unknown variable              admin1 (adm1?)
#> 2  unknown variable             lat (latitude?)
#> 3  unknown variable    longitude.1 (longitude?)
#> 4  unknown variable K_fertilize (K_fertilizer?)
#> 5 unknown variables                        test

check_variables(x, trms, required=TRUE)
#>               check
#> 1  unknown variable
#> 2  unknown variable
#> 3  unknown variable
#> 4  unknown variable
#> 5 unknown variables
#> 6 missing variables
#>                                                                           msg
#> 1                                                              admin1 (adm1?)
#> 2                                                             lat (latitude?)
#> 3                                                    longitude.1 (longitude?)
#> 4                                                 K_fertilize (K_fertilizer?)
#> 5                                                                        test
#> 6 trial_id, on_farm, is_survey, yield_moisture, geo_from_source, K_fertilizer

vocal::check_values(x, trms)
#>           check                                 msg
#> 1  bad datatype dataset_id (numeric, not character)
#> 2 invalid terms                        crop: manioc

check_date(x, "planting_date")
#>   check                              msg
#> 1  date future date(s) in: planting_date

check_datespan(x, "planting_date", "harvest_date")
#>      check                                                   msg
#> 1 datespan 1 harvest_date more than 366 days after planting_date

check_lonlat(x)
#> Cached as: /home/runner/.local/share/.carob/gadm/gadm36_adm0_r1_pk.rds
#>                          check                    msg
#> 1 coordinates in wrong country Cameroon/Côte d'Ivoire